(minor) move the api/http helper to helpers/ directory and adjust the api/interfece imports (since it will be used in other components and not only in api/*)
